+++ docs/src/app.js
@@ -3,6 +3,7 @@ document.addEventListener('DOMContentLoaded', () => {
     delimiters: ['${', '}'],
     el: '#app',
     data: {
+      bpTabActive: 'xs',
       breakpoints: {
         xs: 'mobile',
         sm: 'mobile landscape',
@@ -32,6 +33,8 @@ document.addEventListener('DOMContentLoaded', () => {
       deckOpen: false,
       deckCol: false,
       deckColSize: '4-sm',
+      xSpaceSize: '4',
+      ySpaceSize: '4',
       deckAnimate: false,
       deckAnimation: 'fadeIn',
       toastOrientation: 'opens-top',